Add a spec_helper to run specs against live redis
You can invoke it by running

    bundle exec rspec -r spec_helper_live_redis.rb spec

And it'll connect to a redis instance running on localhost:6379/0, and run the test suite against it. Making sure that the specs are accurate against a live redis instance running redis-rb 3.0.0 (at the time of committing). This is so we can be sure that our specs are accurate compared to what live redis would return, so we can have an accurate fake-redis mocking out live redis.