Re-use instances with the same options
My motivation for doing this is it's more akin to how a live redis object being created works. Connect with the same set of options, get access to the same data in that process. For example, it means you can have multiple blocks of work (think workers processing a series of jobs), all creating new redis connections with the same options and sharing the same values left from the last job processed.