Allow Two clients, one instance, two dbs, at once.
We have to be able to handle two clients with the same host/port accessing different databases at once without overwriting each other. So we store our "data" outside the client instances, in a class level data structure. Client instances access it with a key made up of their host/port, and then select which DB out of the array of them they want. Allows the access we need.