For rspec, clear in the around block instead of the before block
Due to hook order we clear on around instead of before. This is
because if we clear in before, before runs AFTER around does and
therefore overrides any redis based changes we make in the around
hook.

If we clear in around, we can make changes to redis in BOTH the
around and before hooks!